# Speed up `r_lgl_which()`