--- a/src/rlang/vec-lgl.c
+++ b/src/rlang/vec-lgl.c
@@ -6,11 +6,11 @@ r_ssize r_lgl_sum(r_obj* x, bool na_true) {
   }
 
   r_ssize n = r_length(x);
+  r_ssize jump_size = n > 200 ? n / 200 : 1;
 
   r_ssize sum = 0;
   const int* p_x = r_lgl_cbegin(x);
-
-  for (r_ssize i = 0; i < n; ++i) {
+  for (r_ssize i = 0; i < n; i += jump_size) {
     // This can't overflow since `sum` is necessarily smaller or equal
     // to the vector length expressed in `r_ssize`.
     int x_i = p_x[i];
@@ -22,7 +22,14 @@ r_ssize r_lgl_sum(r_obj* x, bool na_true) {
     }
   }
 
-  return sum;
+  if (jump_size == 1) {
+    return(sum);
+  }
+
+  r_ssize guess = 1.2 * sum * jump_size;
+  guess = guess == 0 ? jump_size : guess;
+  guess = guess > n ? n : guess;
+  return(guess);
 }
 
 r_obj* r_lgl_which(r_obj* x, bool na_propagate) {
@@ -42,20 +49,59 @@ r_obj* r_lgl_which(r_obj* x, bool na_propagate) {
   r_obj* which = KEEP(r_alloc_integer(which_n));
   int* p_which = r_int_begin(which);
 
+  if (n == which_n) {
+    for (int i = 0; i < n; ++i) {
+      int elt = p_x[i];
+
+      if (elt) {
+        if (na_propagate && elt == r_globals.na_lgl) {
+          *p_which = r_globals.na_int;
+          ++p_which;
+        } else if (elt != r_globals.na_lgl) {
+          *p_which = i + 1;
+          ++p_which;
+        }
+      }
+    }
+
+    FREE(2);
+    return which;
+  }
+
+  r_ssize size_real = 0;
+  int nprot = 0;
+
   for (int i = 0; i < n; ++i) {
     int elt = p_x[i];
 
     if (elt) {
       if (na_propagate && elt == r_globals.na_lgl) {
+        ++size_real;
+        if (size_real > which_n) {
+          which_n = which_n * 1.3;
+          which = KEEP(r_vec_resize(which, which_n));
+          p_which = r_int_begin(which) + size_real - 1;
+          ++nprot;
+        }
         *p_which = r_globals.na_int;
         ++p_which;
       } else if (elt != r_globals.na_lgl) {
+        ++size_real;
+        if (size_real > which_n) {
+          which_n = which_n * 1.3;
+          which = KEEP(r_vec_resize(which, which_n));
+          p_which = r_int_begin(which) + size_real - 1;
+          ++nprot;
+        }
         *p_which = i + 1;
         ++p_which;
       }
     }
   }
 
-  FREE(1);
-  return which;
+  r_obj* out = KEEP(r_int_resize(which, size_real));
+
+  FREE(2);
+  FREE(nprot);
+  return out;
 }